+++ packages/elements/src/interface.d.ts
@@ -1,2 +1,3 @@
 export * from './components';
+export { config as inoElementsConfig, InoElementsConfig, Config, InoElementsWindow, setupConfig } from './components/config';
 export * from './components/types';